start hike button on watch starts hike function, still bad server error, will fix later
