+++ that-guide/src/components/stopwatch/watch_display/WatchDisplay.js
@@ -2,8 +2,9 @@ import React, { useState } from "react";
 import "./WatchDisplay.css";
 import Timer from "../timer/Timer";
 import WatchButtons from "../watch_buttons/WatchButtons";
+import axios from "axios";
 
-function StopWatch() {
+function StopWatch({ latitude, longitude }) {
   const [isActive, setIsActive] = useState(false);
   const [isPaused, setIsPaused] = useState(true);
   const [time, setTime] = useState(0);
@@ -24,14 +25,44 @@ function StopWatch() {
     };
   }, [isActive, isPaused]);
 
-  const handleStart = () => {
+
+  const [endHike, setEndHike] = useState(null);
+  const [distanceTraveled, setDistanceTraveled] = useState(null);
+  const [speed, setSpeed] = useState(null);
+  const [timeTraveled, setTimeTraveled] = useState(null);
+  const [elevationChange, setElevationChange] = useState(null);
+  const [startLat, setStartLat] = useState(latitude);
+  const [startLong, setStartLong] = useState(longitude);
+  const handleStartHike = (event) => {
+    console.log("hello button");
+    // event.preventDefault();
+    // setStartLat(latitude);
+    // setStartLong(longitude);
     setIsActive(true);
     setIsPaused(false);
-    console.log(
-      "start session post request (sending that data that we do have access to ie. start_location etc), get current location"
-    );
+    axios
+      .post(`https://thatguide.herokuapp.com/map/`, {
+        start_location: {
+          latitude: startLat,
+          longitude: startLong,
+        },
+        end_location: endHike,
+        distance_traveled: distanceTraveled,
+        avg_mph: speed,
+        travel_time: timeTraveled,
+        elevation_gain: elevationChange
+      })
+      .then((res) => {
+        console.log("posted something")
+      })
+  }
 
-  };
+  // const handleStart = () => {
+  //   console.log(
+  //     "start session post request (sending that data that we do have access to ie. start_location etc), get current location"
+  //   );
+
+  // };
 
   const handlePauseResume = () => {
     console.log(`time at pause in milliseconds is ${time}`);
@@ -57,7 +88,7 @@ function StopWatch() {
       <WatchButtons
         active={isActive}
         isPaused={isPaused}
-        handleStart={handleStart}
+        handleStart={handleStartHike}
         handlePauseResume={handlePauseResume}
         handleReset={handleReset}
         handleStop={handleStop}
